backend/requirements.txt
flask>=2.3.0
fastapi>=0.100.0
uvicorn>=0.22.0
gitpython>=3.1.0
bandit>=1.7.5
semgrep>=1.30.0
openai>=1.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.0
redis>=4.5.0
celery>=5.3.0
WeasyPrint>=59.0
pdfkit>=1.0.0
markdown2>=2.4.0
python-multipart>=0.0.6
aiofiles>=23.1.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
